Increase some UI precision

+++ src/devices/device_config.h
@@ -20,7 +20,7 @@ struct DeviceConfiguration {
   minMax<short> crop_x{-10000, 10000};
   minMax<short> crop_y{-10000, 10000};
   minMax<short> crop_z{-10000, 10000};
-  short3 offset;
+  float3 offset;
   float3 rotation_deg{0, 0, 0};
   float scale;
   int sample = 1;
